+++ sequence/sequences/integer/recursive.py
@@ -9,12 +9,10 @@ from sequence.sequences.integer.recursive_generalised_sequences import (
 
 class A000032(LucasSequenceV):
     """Lucas numbers (https://oeis.org/A000032)."""
+    sequence_name = 'Lucas numbers'
 
     def __init__(self):
         super().__init__(p=1, q=-1)
-
-    def __str__(self):
-        return 'Lucas numbers'
 
     def __contains__(self, item: Any) -> bool:
         if item == 1:
@@ -27,12 +25,10 @@ LucasNumbers = A000032
 
 class A000045(LucasSequenceU):
     """Fibonacci numbers (https://oeis.org/A000045)."""
+    sequence_name = 'Fibonacci numbers'
 
     def __init__(self):
         super().__init__(p=1, q=-1)
-
-    def __str__(self):
-        return 'Fibonacci numbers'
 
 
 FibonacciNumbers = A000045
@@ -41,12 +37,10 @@ FibonacciSequence = A000045
 
 class A000073(HighOrderFibonacciNumbers):
     """Tribonacci numbers (https://oeis.org/A000073)."""
+    sequence_name = 'Tribonacci numbers'
 
     def __init__(self):
         super().__init__(order=3)
-
-    def __str__(self):
-        return 'Tribonacci numbers'
 
 
 TribonacciNumbers = A000073
@@ -54,12 +48,10 @@ TribonacciNumbers = A000073
 
 class A000078(HighOrderFibonacciNumbers):
     """Tetranacci numbers (https://oeis.org/A000078)."""
+    sequence_name = 'Tetranacci numbers'
 
     def __init__(self):
         super().__init__(order=4)
-
-    def __str__(self):
-        return 'Tetranacci numbers'
 
 
 TetranacciNumbers = A000078
@@ -67,12 +59,10 @@ TetranacciNumbers = A000078
 
 class A000129(LucasSequenceU):
     """Pell numbers (https://oeis.org/A000129)."""
+    sequence_name = 'Pell numbers'
 
     def __init__(self):
         super().__init__(p=2, q=-1)
-
-    def __str__(self):
-        return 'Pell numbers'
 
 
 PellNumbers = A000129
@@ -81,11 +71,10 @@ LambdaNumbers = A000129
 
 class A001591(HighOrderFibonacciNumbers):
     """Pentanacci numbers (https://oeis.org/A001591)."""
+    sequence_name = 'Pentanacci numbers'
+
     def __init__(self):
         super().__init__(order=5)
-
-    def __str__(self):
-        return 'Pentanacci numbers'
 
 
 PentanacciNumbers = A001591
@@ -93,11 +82,10 @@ PentanacciNumbers = A001591
 
 class A001592(HighOrderFibonacciNumbers):
     """Hexanacci numbers (https://oeis.org/A001592)."""
+    sequence_name = 'Hexanacci numbers'
+
     def __init__(self):
         super().__init__(order=6)
-
-    def __str__(self):
-        return 'Hexanacci numbers'
 
 
 HexanacciNumbers = A001591
@@ -105,11 +93,10 @@ HexanacciNumbers = A001591
 
 class A002203(LucasSequenceV):
     """Companion Pell numbers (https://oeis.org/A002203)."""
+    sequence_name = 'Companion Pell numbers'
+
     def __init__(self):
         super().__init__(p=2, q=-1)
-
-    def __str__(self):
-        return 'Companion Pell numbers'
 
 
 CompanionPellNumbers = A002203
@@ -118,11 +105,10 @@ PellLucasNumbers = A002203
 
 class A079262(HighOrderFibonacciNumbers):
     """Octanacci numbers (https://oeis.org/A079262)."""
+    sequence_name = 'Octanacci numbers'
+
     def __init__(self):
         super().__init__(order=8)
-
-    def __str__(self):
-        return 'Octanacci numbers'
 
 
 OctanacciNumbers = A079262
@@ -130,11 +116,10 @@ OctanacciNumbers = A079262
 
 class A104144(HighOrderFibonacciNumbers):
     """Enneanacci numbers (https://oeis.org/A104144)."""
+    sequence_name = 'Enneanacci numbers'
+
     def __init__(self):
         super().__init__(order=9)
-
-    def __str__(self):
-        return 'Enneanacci numbers'
 
 
 EnneanacciNumebrs = A104144
@@ -142,11 +127,10 @@ EnneanacciNumebrs = A104144
 
 class A122189(HighOrderFibonacciNumbers):
     """Heptanacci numbers (https://oeis.org/A122189)."""
+    sequence_name = 'Heptanacci numbers'
+
     def __init__(self):
         super().__init__(order=7)
-
-    def __str__(self):
-        return 'Heptanacci numbers'
 
 
 HeptanacciNumbers = A122189
@@ -154,12 +138,10 @@ HeptanacciNumbers = A122189
 
 class A214733(LucasSequenceU):
     """Sequence A214733 (https://oeis.org/A214733)."""
+    sequence_name = 'sequence A214733'
 
     def __init__(self):
         super().__init__(p=-1, q=3)
-
-    def __str__(self):
-        return 'sequence A214733'
 
     def __contains__(self, item):
         if item == 0: