+++ sequence/sequences/integer/recursive_generalised_sequences.py
@@ -11,15 +11,13 @@ class HighOrderFibonacciNumbers(MonotonicIncreasing, Recursive):
         Attributes:
             order (int): The order of the Fibonacci sequence.
     """
+    sequence_name = 'High order Fibonacci numbers'
 
     def __init__(self, order: int):
         self.order = validate_positive_integer(integer=order)
         start_terms = self._compute_start_terms()
 
         super().__init__(start_terms=start_terms)
-
-    def __str__(self):
-        return 'High order Fibonacci numbers'
 
     def _compute_start_terms(self) -> Tuple[Any, ...]:
         """
@@ -43,14 +41,12 @@ class LucasSequenceU(MonotonicIncreasing, Recursive):
         p (int): Coefficient for the U_{n+1} term.
         q (int): Coefficient for the U_n term.
     """
+    sequence_name = 'Lucas sequence U_n'
 
     def __init__(self, p: int, q: int):
         super().__init__(start_terms=(0, 1))
         self.p = validate_integer(integer=p)
         self.q = validate_integer(integer=q)
-
-    def __str__(self):
-        return 'Lucas sequence U_n'
 
     def formula(self, terms: Tuple[int, int]) -> Tuple[int, int]:
         return terms[1], self.p * terms[1] - self.q * terms[0]
@@ -67,11 +63,9 @@ class LucasSequenceV(LucasSequenceU):
         p (int): Coefficient for the V_{n+1} term.
         q (int): Coefficient for the V_n term.
     """
+    sequence_name = 'Lucas sequence V_n'
 
     def __init__(self, p: int, q:  int):
         super().__init__(p=p, q=q)
 
         self.start_terms = (2, self.p)
-
-    def __str__(self):
-        return 'Lucas sequence V_n'